back/requirements.txt
asgiref==3.8.1
Django==4.2.14
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
greenlet==3.0.3
importlib-metadata==8.2.0
Markdown==3.6
mychat==0.2.6
pillow==10.4.0
PyJWT==2.8.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.14
PyQt5-sip==12.15.0
python-dotenv==1.0.1
rest-framework-simplejwt==0.0.2
SQLAlchemy==2.0.31
sqlparse==0.5.1
typing-extensions==4.12.2
zipp==3.19.2
